+++ test/unit/Base.mocks.js
@@ -7,7 +7,8 @@ const Mock = function () {
   const sandbox = test.sinon.createSandbox()
 
   const connectionStub = {
-    post: sandbox.stub().usingPromise().resolves()
+    post: sandbox.stub().usingPromise().resolves(),
+    get: sandbox.stub().usingPromise().resolves()
   }
 
   const serviceStubs = {
@@ -31,8 +32,9 @@ const Mock = function () {
       MethodNotAllowed: sandbox.stub().returns(new Error()),
       Forbidden: sandbox.stub().returns(new Error()),
       Conflict: sandbox.stub().returns(new Error()),
-      ClienTimeOut: sandbox.stub().returns(new Error()),
-      ServerUnavailable: sandbox.stub().returns(new Error())
+      ClientTimeOut: sandbox.stub().returns(new Error()),
+      ServerUnavailable: sandbox.stub().returns(new Error()),
+      BadGateway: sandbox.stub().returns(new Error())
     },
     client: {
       Connection: sandbox.stub().returns(connectionStub),